STARTREK: Split up startrek.cpp into more files
